refactor: migrate EpochService mock in epoch service
